requirements.txt
fastapi
uvicorn[standard]
pydantic
pytest
httpx
google-generativeai
python-dotenv
streamlit